DOMapp; Domain Name Search Web App
DOMapp was made to work as a WebView contained app in a mobile device environment, mainly speaking iOS and Android, as well as a web-based application for web-browser use. Requires a back-end data processor for it to work, however the core of the web based application can be adapted for many different uses and functions.

By Luis Gustavo Rodriguez (drlouie).