Cronjobs, State Handlers and User Management
/var_local/ files belong in /var/local/ on a UNIX/Linux machine, or anywhere else for that matter. However, many pointers are looking for this location so why not keep it the same to make it easy on yourself?

Cronjobs are our main way of keeping out data up to date, there's one for each data point. The cronjobs we use are in the 'cronnie' directory, each are usually run every few minutes. Needless to say, these .pl files need to be executable.

We keep the state of a data point, such as the moon, which doesn't change too often and is fairly easy to keep state on, as well as the Department of Homeland Security's Threat Level Advisory system, in one spot where its safe and easily accessed by any means within the system, in this a case a flat file. Sure we use many methods to keep state of the state, but for quick reading we do so in a flat file.

File is nmscontrol are basically controllers for the app, for instance user alert management and secure remote server file transfer connection, etc.