vps-net.com's invisibleHTTP
The invisibleHTTP directory exists parallel to the http facing root directory and is referenced for data we'd rather keep away from prying eyes. It is also a place to work with data away from http areas but close enough, and locally placed, to allow full and proper permissions management for lesser users. 

You will notice .js.nsf files, these are JavaScript files in their pre-obfuscation state. I've included these to allow you to see how I enacted virtual obfuscation of my custom scripts to keep malicious users at bay or bot driven activities at minimal. If you want to see these JavaScript files in their post-obfuscation state just look for them, live, at http://vps-net.com. 

For instance, to find the obfuscated version of CarouselHeadScript.js.nsf you have two call options:

Directly: http://vps-net.com/cross_browser_javascripts/WikiPeeksCore-Wikipedia-1.0.0.js
Through the obfuscator: http://vps-net.com/interface_design_templater.php?p=WPWPHS&q=1

Notice how the obfuscation mechanism routed call adds a comment to the top of the parsed script, /*WikiPeeksWikipediaHead*/? This telling the developer that this is indeed a pre-processed file opposed to a flat file served from the server itself.

The way it works is if an developer is logged into the system their calls through the obfuscation mechanism are first routed through to see if a pre-obfuscation state file for the requested script is available in the invisibleHTTP directory. Meaning, if the .nsf file exists in invisibleHTTP directory this is more than likely due to the fact a developer has placed it there to work on it. It, in turn, is parsed for the developer and only the developer. 

Why? Lets say a developer is writing and checking their code, this is the way to do it actively on the website as it sits without interfering with the general public's use of the site. More so considering the end user will always receive the version of the script living in http://vps-net.com/cross_browser_javascripts/ no matter what type of request they make for the script, flat or otherwise, never anything in invisibleHTTP.

Enjoy my trickery :)